requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Requests==2.32.3
spacy==3.7.5
python-dotenv==1.0.1